requirement.txt
openai
tiktoken
langchain
langchain-community
streamlit
faiss-cpu
python-dotenv
PyPDF2

-e .